Refactor CLI entry point to main.py
Refactor the code to define the main entry point CLI definitions and usage in `main.py`.

* **main.py**
  - Import necessary modules and functions.
  - Define the `cli_entry` function to handle CLI definitions and usage.
  - Replace the `main` function call with the `cli_entry` function call.

* **droplet_setup.py**
  - Remove the `main` function definition.
  - Remove the `cli_entry` function definition.
